+++ src/Parameters.h
@@ -43,7 +43,7 @@ private:
     int INIT_NUM_HELPERS;     //initial number of helpers per group
     double BIAS_FLOAT_BREEDER; //mean of number of groups a floater can visit to try to become a breeder compared to 1 group for helpers
     int MIN_AGE_BECOME_BREEDER; //minimum age for individuals to be able to become breeders
-    int FIXED_GROUP_SIZE;       //in the implementation of no group augmentation, virtual group size for survival for breeder and helpers
+    double FIXED_GROUP_SIZE;       //in the implementation of no group augmentation, virtual group size for survival for breeder and helpers
 
 // Modifiers in survival.
     double m;      //base mortality
@@ -142,7 +142,7 @@ public:
 
     double getBiasFloatBreeder() const;
 
-    int getFixedGroupSize() const;
+    double getFixedGroupSize() const;
 
     int getMinAgeBecomeBreeder() const;
